Move HeaderTests to acceptance testing, a Transformations feat
